refactor: ActionPanelsがhookを直接呼び出すように変更 (step-7)
- ActionPanels が useMyVillageSituation/useVillageSituation/useVillageDebugInfo を直接呼び出し
- ActionPanels が participate 状態 (participateError/onParticipated) を内部に持つ
- ActionPanels が canAction/isLatestDay/currentDay を内部で算出
- SayPanel/SayPreviewArea が useRandomKeywords() を直接呼び出し
- route.tsx から不要な props/state/import を削除
